Speed up TCP data tranfer connection
TCP_QUICKACK is an option that forces an ACK (acknowledgment) packet to be sent for every individual packet received, rather than delaying ACKs for multiple packets. Although this may seem counterintuitive, it can actually improve performance. The reason is that the system-dependent delay in waiting for new packets to arrive can be longer than the time it takes to send an ACK immediately. By using TCP_QUICKACK, we have observed better overall results.