Get Calibration Data on FlexMod enabling only
The fix is needed because we were trying to load the calibration data in the init routine, altough this is not supported by some firmware versions (actually, **no official** FW version is supporting it yet). Because of that, using the GUI v0.3.0 with a tofCam660 as we deliver it  to our customers would not work. This at least partially helps for #20, but further investigations on that topic are needed. Most important: nothing should be done in the init routine that requires a specific firmware version (on any camera).